feat: :sparkles: partition to 3
