+++ makefile
@@ -6,7 +6,7 @@
 #    By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/09/09 14:12:20 by youkim            #+#    #+#              #
-#    Updated: 2021/12/16 11:08:23 by youkim           ###   ########.fr        #
+#    Updated: 2021/12/16 11:52:24 by youkim           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -30,8 +30,8 @@ VSFLAGS  := --show-reachable=yes --error-limit=no --gen-suppressions=all \
 			# --log-file=./mlx.supp
 
 HGEN     := hgen #../hgen/src/run.py
-TPARAM   := 2 16 4 6 3 14 1 11 15 5 7 12 10 9 8 13
-#$(shell ruby -e "puts (1..16).to_a.shuffle.join(' ')")
+TPARAM   := $(shell ruby -e "puts (1..16).to_a.shuffle.join(' ')")
+#2 16 4 6 3 14 1 11 15 5 7 12 10 9 8 13
 TEST	 := ./$(NAME) $(TPARAM)
 # ===== Packages =====
 PKGS     := engine utils quicksort
