+++ src/quicksort/number_utils.c
@@ -6,7 +6,7 @@
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 09:51:43 by youkim            #+#    #+#             */
-/*   Updated: 2021/12/16 11:40:55 by youkim           ###   ########.fr       */
+/*   Updated: 2021/12/16 11:53:23 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,18 +35,8 @@ void	set_pivot(t_engine *e, t_flag from, int size, int pivot[2])
 	set_dequnpack(e, from, size, arr);
 	yqsort(size, arr);
 	i = -1;
-	pivot[LO] = ERR;
-	pivot[HI] = ERR;
-	while (++i < size)
-	{
-		if (curs->num == arr[size / 3])
-			pivot[LO] = i;
-		if (curs->num == arr[size / 3 * 2])
-			pivot[HI] = i;
-		curs = curs->lower;
-	}
-	yassert(pivot[LO] != ERR && pivot[HI] != ERR,
-		"set_pivot: could not find pivots");
+	pivot[LO] = arr[size / 3];
+	pivot[HI] = arr[size / 3 * 2];
 	free(arr);
 }
